requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
altair==6.0.0
anyio==4.3.0
apache-airflow==2.9.1
apache-airflow-providers-common-io==1.3.1
apache-airflow-providers-common-sql==1.12.0
apache-airflow-providers-fab==1.0.4
apache-airflow-providers-ftp==3.8.0
apache-airflow-providers-http==4.10.1
apache-airflow-providers-imap==3.5.0
apache-airflow-providers-smtp==1.6.1
apache-airflow-providers-sqlite==3.7.1
apispec==6.6.1
argcomplete==3.3.0
asgiref==3.8.1
attrs==23.2.0
Babel==2.14.0
blinker==1.8.1
cachelib==0.9.0
cachetools==7.0.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clickclick==20.10.2
cmdstanpy==1.3.0
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.3.3
cron-descriptor==1.4.3
croniter==2.0.5
cryptography==41.0.7
cycler==0.12.1
Deprecated==1.2.14
dill==0.3.8
dnspython==2.6.1
docutils==0.16
email_validator==2.1.1
Faker==40.11.0
Flask==2.2.5
Flask-AppBuilder==4.4.1
Flask-Babel==2.0.0
Flask-Caching==2.2.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.6.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
fonttools==4.62.1
frozenlist==1.4.1
fsspec==2023.12.2
gitdb==4.0.12
GitPython==3.1.46
google-re2==1.1.20240501
googleapis-common-protos==1.63.0
grpcio==1.63.0
gunicorn==22.0.0
h11==0.14.0
holidays==0.93
httpcore==0.16.3
httpx==0.23.3
idna==3.7
importlib-metadata==6.11.0
importlib_resources==6.4.0
inflection==0.5.1
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.5.3
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.5.0
lazy-object-proxy==1.10.0
limits==3.11.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.10.8
mdit-py-plugins==0.4.0
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.2.0
multidict==6.0.5
narwhals==2.18.0
numpy==2.4.3
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
ordered-set==4.1.0
packaging==24.0
pandas==2.3.3
pathspec==0.12.1
pendulum==3.0.0
pillow==12.1.1
plotly==6.6.0
pluggy==1.5.0
prison==0.2.1
prophet==1.3.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.11
pyarrow==23.0.1
pycparser==2.22
pydeck==0.9.1
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.3.2
pytest==9.0.2
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
referencing==0.31.1
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.7.1
rich-argparse==1.4.0
rpds-py==0.18.0
scikit-learn==1.8.0
scipy==1.17.1
setproctitle==1.3.3
setuptools==66.1.1
six==1.16.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==1.4.52
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.0
stanio==0.5.1
streamlit==1.55.0
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.6.0
time-machine==2.14.1
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2024.1
uc-micro-py==1.0.3
unicodecsv==0.14.1
universal_pathlib==0.2.2
urllib3==2.2.1
Werkzeug==2.2.3
wheel==0.46.3
wirerope==0.4.7
wrapt==1.16.0
WTForms==3.1.2
yarl==1.9.4
zipp==3.18.1
